+++ cirq-core/cirq/experiments/randomized_and_cross_entropy_benchmarking/grid_parallel_two_qubit_xeb.py
@@ -15,7 +15,7 @@
 
 This module keeps GridParallelXEBMetadata around for JSON backwards compatibility.
 """
-import dataclasses
+from dataclasses import dataclass
 from typing import Optional, Sequence, TYPE_CHECKING
 from cirq.experiments.random_quantum_circuit_generation import GridInteractionLayer
 from cirq import protocols
@@ -30,11 +30,18 @@ LAYER_C = GridInteractionLayer(col_offset=1, vertical=False, stagger=True)
 LAYER_D = GridInteractionLayer(col_offset=0, vertical=False, stagger=True)
 
 
-@dataclasses.dataclass
+@dataclass(frozen=True, repr=False)
 class GridParallelXEBMetadata:
     """Metadata for a grid parallel XEB experiment.
+
     Attributes:
-        data_collection_id: The data collection ID of the experiment.
+        qubits: The qubits used in the experiment.
+        two_qubit_gate: The entangling gate.
+        num_circuits: Number of circuits.
+        repetitions: Number of repetitions.
+        cycles: Sequence of number of cycles.
+        layers: Grid interaction layers.
+        seed: pseudo-random number generator seed.
     """
 
     qubits: Sequence['cirq.Qid']
@@ -50,8 +57,8 @@ class GridParallelXEBMetadata:
 
     def __repr__(self) -> str:
         return (
-            'cirq.experiments.grid_parallel_two_qubit_xeb.'
-            'GridParallelXEBMetadata('
+            'cirq.experiments.randomized_and_cross_entropy_benchmarking.'
+            'grid_parallel_two_qubit_xeb.GridParallelXEBMetadata('
             f'qubits={self.qubits!r}, '
             f'two_qubit_gate={self.two_qubit_gate!r}, '
             f'num_circuits={self.num_circuits!r}, '
