+++ cirq-core/cirq/experiments/randomized_and_cross_entropy_benchmarking/two_qubit_xeb_test.py
@@ -24,7 +24,9 @@ import pandas as pd
 import pytest
 
 import cirq
-from cirq.experiments.qubit_characterizations import ParallelRandomizedBenchmarkingResult
+from cirq.experiments.randomized_and_cross_entropy_benchmarking.randomized_benchmarking import (
+    ParallelRandomizedBenchmarkingResult,
+)
 
 
 def _manhattan_distance(qubit1: 'cirq.GridQubit', qubit2: 'cirq.GridQubit') -> int:
