+++ cirq-core/cirq/protocols/json_test_data/GridParallelXEBMetadata.repr
@@ -1,1 +1,1 @@
-[cirq.experiments.grid_parallel_two_qubit_xeb.GridParallelXEBMetadata(qubits=[cirq.GridQubit(0, 0), cirq.GridQubit(0, 1), cirq.GridQubit(1, 0), cirq.GridQubit(1, 1)], two_qubit_gate=cirq.ISWAP, num_circuits=10, repetitions=10000, cycles=[2, 4, 6, 8, 10], layers=[cirq.experiments.GridInteractionLayer(col_offset=0, vertical=True, stagger=True), cirq.experiments.GridInteractionLayer(col_offset=1, vertical=True, stagger=True)], seed=1234), cirq.experiments.grid_parallel_two_qubit_xeb.GridParallelXEBMetadata(qubits=[cirq.GridQubit(0, 0), cirq.GridQubit(0, 1), cirq.GridQubit(0, 2), cirq.GridQubit(1, 0), cirq.GridQubit(1, 1), cirq.GridQubit(1, 2)], two_qubit_gate=cirq.CZ, num_circuits=50, repetitions=1000, cycles=[3, 6, 9, 12], layers=[cirq.experiments.GridInteractionLayer(col_offset=1, vertical=False, stagger=True), cirq.experiments.GridInteractionLayer(col_offset=0, vertical=False, stagger=True)], seed=None)]
+[cirq.experiments.randomized_and_cross_entropy_benchmarking.grid_parallel_two_qubit_xeb.GridParallelXEBMetadata(qubits=[cirq.GridQubit(0, 0), cirq.GridQubit(0, 1), cirq.GridQubit(1, 0), cirq.GridQubit(1, 1)], two_qubit_gate=cirq.ISWAP, num_circuits=10, repetitions=10000, cycles=[2, 4, 6, 8, 10], layers=[cirq.experiments.GridInteractionLayer(col_offset=0, vertical=True, stagger=True), cirq.experiments.GridInteractionLayer(col_offset=1, vertical=True, stagger=True)], seed=1234), cirq.experiments.randomized_and_cross_entropy_benchmarking.grid_parallel_two_qubit_xeb.GridParallelXEBMetadata(qubits=[cirq.GridQubit(0, 0), cirq.GridQubit(0, 1), cirq.GridQubit(0, 2), cirq.GridQubit(1, 0), cirq.GridQubit(1, 1), cirq.GridQubit(1, 2)], two_qubit_gate=cirq.CZ, num_circuits=50, repetitions=1000, cycles=[3, 6, 9, 12], layers=[cirq.experiments.GridInteractionLayer(col_offset=1, vertical=False, stagger=True), cirq.experiments.GridInteractionLayer(col_offset=0, vertical=False, stagger=True)], seed=None)]
